+++ packages/studio-base/src/panels/Plot/builders/IndexDatasetsBuilder.test.ts
@@ -2,7 +2,6 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { PlotConfig } from "@foxglove/studio-base/panels/Plot/types";
 import {
   MessageBlock,
   PlayerPresence,
@@ -11,6 +10,7 @@ import {
 } from "@foxglove/studio-base/players/types";
 
 import { IndexDatasetsBuilder } from "./IndexDatasetsBuilder";
+import { PlotConfig } from "../config";
 
 function buildPlotConfig(override: Partial<PlotConfig>): PlotConfig {
   return {
