+++ packages/studio-base/src/panels/Plot/builders/IndexDatasetsBuilder.ts
@@ -17,10 +17,9 @@ import { getContrastColor, getLineColor } from "@foxglove/studio-base/util/plotC
 
 import { CsvDataset, GetViewportDatasetsResult, IDatasetsBuilder } from "./IDatasetsBuilder";
 import { Dataset } from "../ChartRenderer";
+import { isReferenceLinePlotPathType, PlotConfig } from "../config";
 import { getChartValue, isChartValue, Datum } from "../datum";
-import { isReferenceLinePlotPathType } from "../internalTypes";
 import { mathFunctions } from "../mathFunctions";
-import { PlotConfig } from "../types";
 
 type DatumWithReceiveTime = Datum & {
   receiveTime: Time;
