+++ packages/studio-base/src/panels/Plot/config.test.ts
@@ -2,7 +2,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { isReferenceLinePlotPathType } from "./internalTypes";
+import { isReferenceLinePlotPathType } from "./config";
 
 describe("isReferenceLinePlotPathType", () => {
   it.each(["0", "1.2", "1e6"])("returns true for '%s'", (value) => {
