+++ packages/studio-base/src/panels/Plot/index.stories.tsx
@@ -20,9 +20,9 @@ import { makeStyles } from "tss-react/mui";
 import PanelSetup, { Fixture, triggerWheel } from "@foxglove/studio-base/stories/PanelSetup";
 import delay from "@foxglove/studio-base/util/delay";
 
+import { PlotConfig } from "./config";
 import Plot from "./index";
 import { fixture } from "./storyFixtures";
-import { PlotConfig } from "./types";
 
 const withEndTime = (testFixture: Fixture, endTime: any) => ({
   ...testFixture,
